В код внесены изменения. Добавлено разнесение классов по файлам заголовков
